Pivot updated: sign of movement is introduced (random +- 90)
Matrix.h: minor improvement

MoveBack has CP procedure -> updating node position and polymer
distributions in rejected states by Metropolis

Test are passed.

+++ src/cleng/matrix.h
@@ -7,8 +7,8 @@
 #include <stdexcept>
 #include "nodes/point.h"
 
-template <class T>
-class Matrix{
+template<class T>
+class Matrix {
 private:
     std::vector<std::vector<T> > array;
     int height;
@@ -16,61 +16,93 @@ private:
 
 public:
     Matrix<T>(int height, int width);
+
     explicit Matrix<T>(std::vector<std::vector<T> > const &array);
+
     Matrix<T>();
 
     int getHeight() const;
+
     int getWidth() const;
 
-    Matrix<T> add(const Matrix<T>& m) const;
-    Matrix<T> subtract(const Matrix<T>& m) const;
-    Matrix<T> multiply(const Matrix<T>& m) const;
-    Matrix<T> dot(const Matrix<T>& m) const;
+    Matrix<T> add(const Matrix<T> &m) const;
+
+    Matrix<T> subtract(const Matrix<T> &m) const;
+
+    Matrix<T> multiply(const Matrix<T> &m) const;
+
+    Matrix<T> dot(const Matrix<T> &m) const;
+
     Point dot(Point point) const;
+
     Matrix<T> negate() const;
+
     Matrix<T> transpose() const;
-    Matrix<T> multiply(const T& value) const;
-    Matrix<T> divide(const T& value) const;
+
+    Matrix<T> multiply(const T &value) const;
+
+    Matrix<T> divide(const T &value) const;
 
     Matrix<T> applyFunction(T (*function)(T)) const;
+
     Matrix<T> subMat(int startH, int startW, int h, int w) const;
 
-    void fill(const T& value);
-    void put(int h, int w, const T& value);
+    void fill(const T &value);
+
+    void put(int h, int w, const T &value);
+
     T get(int h, int w) const;
 
     void print(std::ostream &flux) const;
 
-    bool operator==(const Matrix<T>& m);
-    bool operator!=(const Matrix<T>& m);
-    Matrix<T> operator+=(const Matrix<T>& m);
-    Matrix<T> operator-=(const Matrix<T>& m);
-    Matrix<T> operator*=(const Matrix<T>& m);
+    bool operator==(const Matrix<T> &m);
+
+    bool operator!=(const Matrix<T> &m);
+
+    Matrix<T> operator+=(const Matrix<T> &m);
+
+    Matrix<T> operator-=(const Matrix<T> &m);
+
+    Matrix<T> operator*=(const Matrix<T> &m);
+
     Matrix<T> operator*=(const T &m);
+
     Matrix<T> operator/=(const T &m);
-    T& operator()(int y, int x);
+
+    T &operator()(int y, int x);
 };
 
-template <class T> Matrix<T> operator+(const Matrix<T>& a, const Matrix<T>& b);
-template <class T> Matrix<T> operator-(const Matrix<T>& a, const Matrix<T>& b);
-template <class T> Matrix<T> operator*(const Matrix<T>& a, const Matrix<T>& b);
-template <class T> Matrix<T> operator*(const T &b, const Matrix<T>& a);
-template <class T> Matrix<T> operator/(const Matrix<T>& a, const T &b);
-template <class T> std::ostream& operator<<(std::ostream &flux, const Matrix<T>& m);
+template<class T>
+Matrix<T> operator+(const Matrix<T> &a, const Matrix<T> &b);
+
+template<class T>
+Matrix<T> operator-(const Matrix<T> &a, const Matrix<T> &b);
+
+template<class T>
+Matrix<T> operator*(const Matrix<T> &a, const Matrix<T> &b);
+
+template<class T>
+Matrix<T> operator*(const T &b, const Matrix<T> &a);
+
+template<class T>
+Matrix<T> operator/(const Matrix<T> &a, const T &b);
+
+template<class T>
+std::ostream &operator<<(std::ostream &flux, const Matrix<T> &m);
 
 #endif //NAMICS_MATRIX_H
 
 
-template <class T>
-Matrix<T>::Matrix(int height, int width){
+template<class T>
+Matrix<T>::Matrix(int height, int width) {
     this->height = height;
     this->width = width;
     this->array = std::vector<std::vector<T> >(height, std::vector<T>(width));
 }
 
-template <class T>
-Matrix<T>::Matrix(std::vector<std::vector<T> > const &array){
-    if(array.size()==0)
+template<class T>
+Matrix<T>::Matrix(std::vector<std::vector<T> > const &array) {
+    if (array.size() == 0)
         throw std::invalid_argument("Size of array must be greater than 0.");
 
     this->height = array.size();
@@ -78,52 +110,52 @@ Matrix<T>::Matrix(std::vector<std::vector<T> > const &array){
     this->array = array;
 }
 
-template <class T>
-Matrix<T>::Matrix(){
+template<class T>
+Matrix<T>::Matrix() {
     height = 0;
     width = 0;
 }
 
-template <class T>
-int Matrix<T>::getHeight() const{
+template<class T>
+int Matrix<T>::getHeight() const {
     return height;
 }
 
-template <class T>
-int Matrix<T>::getWidth() const{
+template<class T>
+int Matrix<T>::getWidth() const {
     return width;
 }
 
-template <class T>
-void Matrix<T>::fill(const T& value){
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+template<class T>
+void Matrix<T>::fill(const T &value) {
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             array[i][j] = value;
         }
     }
 }
 
-template <class T>
-void Matrix<T>::put(int h, int w, const T& value){
-    if(!(h>=0 && h<height && w>=0 && w<width))
+template<class T>
+void Matrix<T>::put(int h, int w, const T &value) {
+    if (!(h >= 0 && h < height && w >= 0 && w < width))
         throw std::invalid_argument("Index out of bounds.");
 
     array[h][w] = value;
 }
 
-template <class T>
-T Matrix<T>::get(int h, int w) const{
-    if(!(h>=0 && h<height && w>=0 && w<width))
+template<class T>
+T Matrix<T>::get(int h, int w) const {
+    if (!(h >= 0 && h < height && w >= 0 && w < width))
         throw std::invalid_argument("Index out of bounds.");
 
     return array[h][w];
 }
 
-template <class T>
-Matrix<T> Matrix<T>::multiply(const T& value) const{
+template<class T>
+Matrix<T> Matrix<T>::multiply(const T &value) const {
     Matrix result(array);
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             result.array[i][j] *= value;
         }
     }
@@ -131,11 +163,11 @@ Matrix<T> Matrix<T>::multiply(const T& value) const{
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::divide(const T& value) const{
+template<class T>
+Matrix<T> Matrix<T>::divide(const T &value) const {
     Matrix result(array);
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             result.array[i][j] /= value;
         }
     }
@@ -143,14 +175,14 @@ Matrix<T> Matrix<T>::divide(const T& value) const{
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::add(const Matrix& m) const{
-    if(!(height==m.height && width==m.width))
+template<class T>
+Matrix<T> Matrix<T>::add(const Matrix &m) const {
+    if (!(height == m.height && width == m.width))
         throw std::invalid_argument("Matrix dimension must be the same.");
 
     Matrix result(height, width);
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             result.array[i][j] = array[i][j] + m.array[i][j];
         }
     }
@@ -158,29 +190,29 @@ Matrix<T> Matrix<T>::add(const Matrix& m) const{
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::subtract(const Matrix& m) const{
-    if(!(height==m.height && width==m.width))
+template<class T>
+Matrix<T> Matrix<T>::subtract(const Matrix &m) const {
+    if (!(height == m.height && width == m.width))
         throw std::invalid_argument("Matrix dimension must be the same.");
 
     Matrix result(height, width);
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             result.array[i][j] = array[i][j] - m.array[i][j];
         }
     }
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::multiply(const Matrix& m) const{
-    if(!(height==m.height && width==m.width))
+template<class T>
+Matrix<T> Matrix<T>::multiply(const Matrix &m) const {
+    if (!(height == m.height && width == m.width))
         throw std::invalid_argument("Matrix dimension must be the same.");
 
     Matrix result(height, width);
 
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             result.array[i][j] = array[i][j] * m.array[i][j];
         }
     }
@@ -190,66 +222,66 @@ Matrix<T> Matrix<T>::multiply(const Matrix& m) const{
 template<class T>
 Point Matrix<T>::dot(Point point) const {
     Point result;
-    for (int i=0 ; i<height ; i++){
-        for (int h=0 ; h<width ; h++){
+    for (int i = 0; i < height; i++) {
+        for (int h = 0; h < width; h++) {
             result[i] += array[i][h] * point[h];
         }
     }
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::dot(const Matrix& m) const{
-    if(width != m.height)
+template<class T>
+Matrix<T> Matrix<T>::dot(const Matrix &m) const {
+    if (width != m.height)
         throw std::invalid_argument("Dot product not compatible.");
 
-    T w=0;
+    T w = 0;
     int mwidth = m.width;
 
     Matrix<T> result(height, mwidth);
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<mwidth ; j++){
-            for (int h=0 ; h<width ; h++){
-                w += array[i][h]*m.array[h][j];
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < mwidth; j++) {
+            for (int h = 0; h < width; h++) {
+                w += array[i][h] * m.array[h][j];
             }
             result.array[i][j] = w;
-            w=0;
+            w = 0;
         }
     }
 
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::negate() const{
+template<class T>
+Matrix<T> Matrix<T>::negate() const {
     Matrix<T> result(width, height);
 
-    for (int i=0 ; i<width ; i++){
-        for (int j=0 ; j<height ; j++){
+    for (int i = 0; i < width; i++) {
+        for (int j = 0; j < height; j++) {
             result.array[i][j] = -array[i][j];
         }
     }
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::transpose() const{
+template<class T>
+Matrix<T> Matrix<T>::transpose() const {
     Matrix<T> result(width, height);
 
-    for (int i=0 ; i<width ; i++){
-        for (int j=0 ; j<height ; j++){
+    for (int i = 0; i < width; i++) {
+        for (int j = 0; j < height; j++) {
             result.array[i][j] = array[j][i];
         }
     }
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::applyFunction(T (*function)(T)) const{
+template<class T>
+Matrix<T> Matrix<T>::applyFunction(T (*function)(T)) const {
     Matrix<T> result(height, width);
 
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             result.array[i][j] = (*function)(array[i][j]);
         }
     }
@@ -257,40 +289,40 @@ Matrix<T> Matrix<T>::applyFunction(T (*function)(T)) const{
     return result;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::subMat(int startH, int startW, int h, int w) const{
-    if(!(startH>=0 && startH+h<=height && startW>=0 && startW+w<=width))
+template<class T>
+Matrix<T> Matrix<T>::subMat(int startH, int startW, int h, int w) const {
+    if (!(startH >= 0 && startH + h <= height && startW >= 0 && startW + w <= width))
         throw std::invalid_argument("Index out of bounds");
 
-    Matrix<T> result(h,w);
-    for (int i=startH ; i<startH+h ; i++){
-        for (int j=startW ; j<startW+w ; j++){
-            result.array[i-startH][j-startW] = array[i][j];
+    Matrix<T> result(h, w);
+    for (int i = startH; i < startH + h; i++) {
+        for (int j = startW; j < startW + w; j++) {
+            result.array[i - startH][j - startW] = array[i][j];
         }
     }
     return result;
 }
 
-template <class T>
-void Matrix<T>::print(std::ostream &flux) const{
+template<class T>
+void Matrix<T>::print(std::ostream &flux) const {
     int maxLength[width] = {};
     std::stringstream ss;
 
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             ss << array[i][j];
-            if(maxLength[j] < ss.str().size()){
+            if (maxLength[j] < (int) ss.str().size()) {
                 maxLength[j] = ss.str().size();
             }
             ss.str(std::string());
         }
     }
 
-    for (int i=0 ; i<height ; i++){
-        for (int j=0 ; j<width ; j++){
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
             flux << array[i][j];
             ss << array[i][j];
-            for (int k=0 ; k<maxLength[j]-ss.str().size()+1 ; k++){
+            for (int k = 0; k < int(maxLength[j] - ss.str().size() + 1); k++) {
                 flux << " ";
             }
             ss.str(std::string());
@@ -299,12 +331,12 @@ void Matrix<T>::print(std::ostream &flux) const{
     }
 }
 
-template <class T>
-bool Matrix<T>::operator==(const Matrix& m){
-    if(height==m.height && width==m.width){
-        for (int i=0 ; i<height ; i++){
-            for (int j=0 ; j<width ; j++){
-                if(array[i][j]!=m.array[i][j]){
+template<class T>
+bool Matrix<T>::operator==(const Matrix &m) {
+    if (height == m.height && width == m.width) {
+        for (int i = 0; i < height; i++) {
+            for (int j = 0; j < width; j++) {
+                if (array[i][j] != m.array[i][j]) {
                     return false;
                 }
             }
@@ -314,74 +346,75 @@ bool Matrix<T>::operator==(const Matrix& m){
     return false;
 }
 
-template <class T>
-bool Matrix<T>::operator!=(const Matrix& m){
+template<class T>
+bool Matrix<T>::operator!=(const Matrix &m) {
     return !operator==(m);
 }
 
-template <class T>
-Matrix<T> Matrix<T>::operator+=(const Matrix& m){
+template<class T>
+Matrix<T> Matrix<T>::operator+=(const Matrix &m) {
     this->array = add(m).array;
     return *this;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::operator-=(const Matrix& m){
+template<class T>
+Matrix<T> Matrix<T>::operator-=(const Matrix &m) {
     this->array = subtract(m).array;
     return *this;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::operator*=(const Matrix& m){
+template<class T>
+Matrix<T> Matrix<T>::operator*=(const Matrix &m) {
     this->array = multiply(m).array;
     return *this;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::operator*=(const T &m){
+template<class T>
+Matrix<T> Matrix<T>::operator*=(const T &m) {
     *this = this->multiply(m);
     return *this;
 }
 
-template <class T>
-Matrix<T> Matrix<T>::operator/=(const T &m){
+template<class T>
+Matrix<T> Matrix<T>::operator/=(const T &m) {
     *this = this->divide(m);
     return *this;
 }
 
-template <class T>
-T& Matrix<T>::operator()(int y, int x){
-    if(!(y>=0 && y<height && x>=0 && x<width))
+template<class T>
+T &Matrix<T>::operator()(int y, int x) {
+    if (!(y >= 0 && y < height && x >= 0 && x < width))
         throw std::invalid_argument("Index out of bounds.");
     return array[y][x];
 }
 
-template <class T>
-Matrix<T> operator+(const Matrix<T>& a, const Matrix<T>& b){
+template<class T>
+Matrix<T> operator+(const Matrix<T> &a, const Matrix<T> &b) {
     return a.add(b);
 }
 
-template <class T>
-Matrix<T> operator-(const Matrix<T>& a, const Matrix<T>& b){
+template<class T>
+Matrix<T> operator-(const Matrix<T> &a, const Matrix<T> &b) {
     return a.subtract(b);
 }
 
-template <class T>
-Matrix<T> operator*(const Matrix<T>& a, const Matrix<T>& b){
+template<class T>
+Matrix<T> operator*(const Matrix<T> &a, const Matrix<T> &b) {
     return a.multiply(b);
 }
 
-template <class T>
-Matrix<T> operator*(const T &b, const Matrix<T>& a){
+template<class T>
+Matrix<T> operator*(const T &b, const Matrix<T> &a) {
     return a.multiply(b);
 }
-template <class T>
-Matrix<T> operator/(const Matrix<T>& a, const T &b){
+
+template<class T>
+Matrix<T> operator/(const Matrix<T> &a, const T &b) {
     return a.divide(b);
 }
 
-template <class T>
-std::ostream& operator<<(std::ostream &flux, const Matrix<T>& m){
+template<class T>
+std::ostream &operator<<(std::ostream &flux, const Matrix<T> &m) {
     m.print(flux);
     return flux;
 }
